Rewrite hover info (#456)
Simplify the computation of type-on-hover information. Previously we created intermediate info datastructures which were then turned into Goto and Hover types required by the LSP server. We now skip this intermediate data structure and generate the LSP types directly. The main reason is that the info structs essentially only reproduced a subset of the information available in the AST.